==> code_raw_knitted/01.05_WIMAS_formatting.Rmd <==
---
title: "WIMAS Data Janitoring"
author: "Jill Deines"
date: "October 9, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Extract Anthony's cleaned WIMAS data from geodatabases and combine into one master file. Mwhahahaha.

Secondary goal: Get familiar with the `sf` package, a new vector GIS package

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/01.05_WIMASformatting/')
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
library(sf)
library(tidyverse) # for ggplot2, tidyr, dplyr
```

## Dataset Description

**Anthony's cleaned dataset: 1990-2016**

* Raw data location: S:\Users\kendal30\Code\Code_Projects\2016\Kansas_WIMAS_Processing
* Output geodatabase: 'S:/Users/kendal30/Project_Files/2017/WIMAS_CHP/Kansas_WIMAS_2016.gdb'

Anthony's output has points clipped to the HPA with the following variables separated into different point shapefiles within a geodatabase in ArcGIS:

* irrigated acres
* irrigated crops
* irrigated depth
* irrigated volume
* irrigation systems
* water rights
* points of diversion

## Load datasets
Here, I load the 5 layers of data for points of diversion produced through Anthony's cleaning: acres, crops, depth, volume, and systems. For now, I ignore the water rights

```{r loadData}
# arcgis geodatabase
gdbfile <- 'S:/Users/kendal30/Project_Files/2017/WIMAS_CHP/Kansas_WIMAS_2016.gdb'

# get geodatabase layer names
st_layers(gdbfile)

# make a df of layer names and desired column names
layerinfo <- data.frame(layername = c('HPA_irrigation_systems','HPA_irrigated_crops',
                                      'HPA_irrigated_volume','HPA_irrigated_depth',
                                      'HPA_irrigated_acres'),
                        columnname = c('system','crop','volume','depth','acres'))

# load similarly formatted data alyers
longlayers <- list()
for (i in 1:nrow(layerinfo)) {
  # load point layer
  widedf0 <- st_read(dsn=gdbfile, layer = layerinfo$layername[i], quiet=TRUE)
  # remove spatial info (retain lat/long column)
  widedf <- as.data.frame(widedf0)
  # drop junk
  lesswide <- widedf %>% select(-c(index, PDIV_ID_1, Shape, F_99))
  # longify
  long <- gather_(lesswide, key = 'year', value = layerinfo$columnname[i], paste0('F', 1990:2016))
  # numeric year
  long$year <- as.numeric(substr(long$year, 2, 5))
  # remove lat/long columns in all but 1 dataset
  if (i > 1){
    long <- long %>% select(-c(LATITUDE,LONGITUDE))
  }
  # store in list
  longlayers[[layerinfo$columnname[i]]] <- long
}

# merge list of dataframes by year and pid
longdata <- Reduce(function(x,y)merge(x,y, by = c('year','PDIV_ID')), longlayers)
  
# 
# # and load water rights, which has almost 8000 more points
rights <- st_read(dsn=gdbfile, layer = 'HPA_water_rights', quiet=TRUE)

# how many are G sourced?
nrow(rights[rights$source_of_supply == 'G',])

#ptsdiv <- st_read(dsn=gdbfile, layer = 'HPA_groundwater_points_of_diversion', quiet=TRUE)
```

There are still 45,638 groundwater rights, which is more than then 38013 points of groundwater diversion. There also isn't a clear linked field between the water rights data (wr_ID) and the ponts of diversion (PDIV_ID). For now, I will ignore the water rights.

## Export dataset for future use

```{r export, eval=FALSE}
outdir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/wellData/WIMAS_AnthonyCleaned_JillFormatted'
write.csv(longdata, row.names = FALSE,
          file = paste0(outdir,'/WIMAS_1990-2016_gwPts_ksHpa.csv'))
```

## Export a shapefile too

```{r exportShape, eval=FALSE}
# reload data
outdir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/wellData/WIMAS_AnthonyCleaned_JillFormatted'
longdata <- read.csv(paste0(outdir,'/WIMAS_1990-2016_gwPts_ksHpa.csv'))

# spatialize, in NAD 27
longpoints <- st_as_sf(longdata, coords=c('LONGITUDE','LATITUDE'), crs = 4267)

# plot
test <- longpoints[longpoints$year == 2000,]
plot(test[1])

# export
write_sf(longpoints, paste0(outdir,'/WIMAS_1990-2016_gwPts_ksHpa.shp'))
```
