==> code_raw_knitted/01.07_WIMAS_sheridanChanges.Rmd <==
---
title: "WIMAS changes in use"
author: "Jill Deines"
date: "December 9, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Quantify changes in pumping during LEMA

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/01.07_wimas_changes/',
                      cache = FALSE)
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
library(sf)
library(tidyverse)
library(xtable)
library(lme4)
library(salustools)
```

## Load data files
Includes formatted data from 01.05 and boundary shapefiles of interest

```{r loadData}
#setwd("C:/Users/adeines/BE_adeines/Personal/Wells for Jill")
# directories
#gisDir <- 'C:/Users/adeines/BE_adeines/Personal/Wells for Jill/lema'
#wellDir <- 'C:/Users/adeines/BE_adeines/Personal/Wells for Jill/lema'

gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries'
wellDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/wellData/WIMAS_AnthonyCleaned_JillFormatted'
precipDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/weather/weatherCheckSheridan'


# well data
wellsLema2 <- read.csv(paste0(wellDir, '/WIMAS_Sheridan_1990-2016.csv'))


prepost <- wellsLema2 %>%
  filter(year > 2001) %>%
  group_by(LEMA) %>%
  summarize(meanVol = mean(volume, na.rm=T), 
            meanAcres = mean(acres, na.rm=T),
            meanDepth = mean(depth, na.rm=T))

longg <- gather(prepost, key = dataset, value = value, meanVol:meanAcres)

ggplot(longg, aes(x=LEMA, y = value)) + 
  geom_bar(stat='identity') + facet_wrap(~dataset)

# plot changes with distributions
plotme <- wellsLema2 %>%
  filter(year > 2001) %>%
  group_by(LEMA,year) %>%
  summarize(totVol = mean(volume, na.rm=T), 
            meanAcres = mean(acres, na.rm=T),
            meanDepth = mean(depth, na.rm=T))

ggplot(plotme %>% filter(dataset !=), aes(x=LEMA, y = value)) +
  geom_boxplot() + facet_wrap(~dataset)
```



## Time Series
Is there autocorrelation to worry about?

```{r autocorrelation, eval=FALSE}
#Volume
for (i in unique(wellsLema2$PDIV_ID)){
  well<-wellsLema2$volume[wellsLema2$PDIV_ID%in%i][-c(1:4)]
  if(sum(is.na(well))==0) acf(ts(well))
  readline("enter")
}
# Most wells volumes don't show any appreciable autocorrelation, but some do, which is tedious.
for (i in unique(wellsLema2$PDIV_ID)){
  well<-diff(wellsLema2$volume[wellsLema2$PDIV_ID%in%i])
  if(sum(is.na(well))==0) acf(ts(well))
  readline("enter")
}
# No consistant autocorrelation pattern.

#Depth
for (i in unique(wellsLema2$PDIV_ID)){
  well<-wellsLema2$depth[wellsLema2$PDIV_ID%in%i]
  if(sum(is.na(well))==0) acf(ts(well))
  readline("enter")
}

````

##Mixed Model
```{r mixed_model, dpi = 400, fig.height = 3, fig.width = 4, dev=c('png','pdf')}
wellsLema2$PDIV_ID<-as.factor(wellsLema2$PDIV_ID)
wellsLema2$m3_vol<-wellsLema2$volume*1233.48
wellsLema2$year0<-wellsLema2$year-2012
wellsLema2_trim<-as.data.frame(wellsLema2[wellsLema2$year>=2002,])

lema_lme00<-lmer(m3_vol~year0*LEMA+(1|PDIV_ID),
               data=wellsLema2_trim)

lema_lme0<-lmer(m3_vol~year*LEMA+(1|PDIV_ID),
               data=wellsLema2_trim)

lema_lme0x<-lmer(m3_vol~year+LEMA+(1|PDIV_ID),
               data=wellsLema2_trim)
lema_lme1<-lmer(m3_vol~year+(1|PDIV_ID),
               data=wellsLema2_trim)
AIC(lema_lme0,lema_lme0x,lema_lme1)

ggplot(data=wellsLema2_trim,aes(y=m3_vol/1000,x=year,color=LEMA))+
         geom_point(alpha=.5)+
         geom_segment(x=2012.5,
                      y=predict(lema_lme0,
                         newdata=data.frame(year=2012.5,LEMA="lema"),re.form=NA)/1000,
                      xend=2016.5,
                      yend=predict(lema_lme0,
                         newdata=data.frame(year=2016.5,LEMA="lema"),re.form=NA)/1000,
                      color="indianred",lwd=1.2)+
         geom_segment(x=2002,
                      y=predict(lema_lme0,
                         newdata=data.frame(year=2002,LEMA="preLema"),re.form=NA)/1000,
                      xend=2012.5,
                      yend=predict(lema_lme0,
                         newdata=data.frame(year=2012.5,LEMA="preLema"),re.form=NA)/1000,
                      color="cornflowerblue",lwd=1.2)+
         geom_segment(x=2012.5,
                      y=predict(lema_lme0,
                         newdata=data.frame(year=2012.5,LEMA="preLema"),re.form=NA)/1000,
                      xend=2016.5,
                      yend=predict(lema_lme0,
                         newdata=data.frame(year=2016.5,LEMA="preLema"),re.form=NA)/1000,
                      color="cornflowerblue",linetype='dashed',lwd=1.2)+
  ylab("Volume (1000m^3)")+
  xlab("")+
  theme_nqblack()+
  theme(legend.position = 'none',
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank())

lema_lme00
1-134322.1/(37686.9+134322.1)


qplot(ranef(lema_lme0)$PDIV_ID)

```
I used a linear mixed effects model, with well as a random effect, to test whether the average volume decreased in the LEMA period.  The mixed effect model used in this way is very similar to a repeated-measures ANOVA on wells.  The signifigance of the LEMA effected was evaluated by comparing models with and without the LEMA using Akaike's information Criterian (AIC).  I found that there was infact a signficant decline in volume pumped in the LEMA period, of about 22%, with a very slightly but significant stepper decline in the LEMA period.  This more complicated model serves two purposes: 1) it better handles misses well data, and 2) suggests a way to incorporate precipiation effects on pumping, but the precipiation model is a work in progress.

```{r compare_to_total_annual}
pre_tot_vol<-sum(wellsLema2_trim$volume[wellsLema2_trim$year<=2012],na.rm=TRUE)/11
lema_tot_vol<-sum(wellsLema2_trim$volume[wellsLema2_trim$year>=2013],na.rm=TRUE)/4

lema_tot_vol/pre_tot_vol

````




```{r with_precip}
ppt<-read.csv(paste0(precipDir,'/Sheridan5_gridmet_meanPrecip_1000.csv'))

wellsLema2_ppt<-merge(wellsLema2,ppt[,c("year","mean")],by="year")
names(wellsLema2_ppt)[names(wellsLema2_ppt)=="mean"]<-"mean_ppt"

ggplot(data=wellsLema2_ppt,aes(y=m3_vol,x=mean_ppt,color=year))+
  geom_point(alpha=.3)+geom_smooth(method="lm")
ggplot(data=wellsLema2_ppt,aes(y=depth,x=mean_ppt,color=year))+
  geom_point(alpha=.3)+geom_smooth(method="lm")

wellsLema2_ppt_trim<-as.data.frame(wellsLema2_ppt[wellsLema2_ppt$year>=2006,])


lema_lme0_ppt<-lmer(m3_vol~year+LEMA+mean_ppt+(1|PDIV_ID),
               data=wellsLema2_ppt_trim)
lema_lme1_ppt<-lmer(m3_vol~year+mean_ppt+(1|PDIV_ID),
               data=wellsLema2_ppt_trim)
AIC(lema_lme0,lema_lme1,lema_lme0_ppt,lema_lme1_ppt)
#lema_lme0_ppt is best by a large margine, LEMA & PPT


ggplot(data=wellsLema2_ppt_trim,aes(y=m3_vol,x=year,color=LEMA))+
         geom_point(alpha=.5)+
         geom_abline(slope=coef(summary(lema_lme0_ppt))["year","Estimate"],
                     intercept=sum(coef(summary(lema_lme0_ppt))["(Intercept)",
                                                                "Estimate"]),
                     color="red")+
          geom_abline(slope=coef(summary(lema_lme0_ppt))["year","Estimate"],
                     intercept=sum(coef(summary(lema_lme0_ppt))[-c(2,4),"Estimate"]),
                     color="blue") +
  theme_nqblack() +
    theme(legend.position='none',
        axis.text=element_text(size=12),
        legend.text=element_text(size=10),
        axis.title=element_text(size=13),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())
```
