==> code_raw_knitted/02.20_SALUS_combinations.Rmd <==
---
title: "Salus Combos"
author: "Jill Deines"
date: "November 5, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Create SALUS gridcell combinations: soil-climate-crop-irrigation combos. Output is a RData dataframe with gridcell ID, and columns x2006:2008 of combinations.

Update 12/6/2017: now includes alfalfa, since omitting it creates a lot of gaps in the 11 year crop rotation sequences. New files named 'top7' for corn, wheat, soy, sorghum, fallow, grass, and ALFALFA

NOTE: 11/29/2017 - the last chunk, 'unique', no longer seems to have enough memory to run, but it's no longer needed (finds unique experiment rotations in the full GMD4 buffer, but on an annual basis, not continuous.)

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/02.20_salus_combos/',
                      cache = TRUE, cache.path='cache/02.20_salus_combos/')
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
library(rgdal)
library(raster)
library(tidyverse)
library(sf)
library(stringr)
```

# user parameters

```{r directories}
# years
yearsWanted <- 2006:2016

# root salus data directory
rootDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus'

# soil inputs
soilDir <- 'soil'
soilPolyName <- 'ssurgo_gmd4p_wgs84'

# crop type inputs
cdlDir <- 'cdl/cdl_rasters_lema'
cdlNameSuffix <- '_CDL_gmd4PlusClips_aea_top7.tif'

# weather grid inputs
weatherDir <- 'weather/nldas_Jill/nldas_centroidExtract'
weatherPolyName <- 'NLDAS_clipped_gmd4p'

# irrigation map inputs
irrDir <- 'irrigationMaps'
irrNameSuffix <- '_gmd4plus_AIM-RRB_aea.tif'

# bounding poly
gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries'
boundName <- 'GMD4_plus_aea'

# static grid export names
comboDir <- 'combinationGrids'
comboStaticName <- 'StaticVars_gmd4p_aea30.tif'
comboDynamicName <- '_gmd4p_combos_top7.RData'
comboMasterFile <- 'gmd4p_combos_top7_master.RData'
allUnique <- 'allUnique_gmd4p_top7_2006-2016.csv'
annualUnique <- 'annualUnique_gmd4p_top7_2006-2016.RData'
```


# Static datasets
Soil and climate grid cell don't change over time. First, add these two together.

## load target grid
Each dataset needs to be reprojected to the same raster grid. I used the CDL grid as the target.

```{r makeGrid}
aeaProj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
template <- raster(nrow = 4359, ncol = 7108, crs = aeaProj,
                    xmn = -522870, xmx = -309630, ymn = 1777860, 
                    ymx = 1908630) 
template[] <- 1:ncell(template)
```

## Soil: 
Load soil mukey polygons, append a state code to the mukeys, and sample to the raster grid

```{r gridsoil}
# load 
mukeys <- read_sf(paste0(rootDir,'/',soilDir,'/',soilPolyName,'.shp'))
# how many soil types?
length(unique(mukeys$MUKEY))

# how long is longest mukey - this makes me laugh
maxdigits <- nchar(as.character(max(as.numeric(mukeys$MUKEY))))

# append a state code to each mukey to result in uniform length
fipsKey <- data.frame(state = c('CO','KS','NE'),
                      fips2 = c(8, 20, 31))
fipsKey$fips2 <- fipsKey$fips2 * 1*10^maxdigits

mukeys <- mukeys %>%
  mutate(state = substr(AREASYMBOL, 1,2)) %>% # extract state abbrev
  left_join(fipsKey) %>%                      # join fips
  mutate(MUKEY = as.numeric(MUKEY),
         stateMukey = MUKEY + fips2) %>%       # enhanced key!
  select(-c(SPATIALVER,Shape_Leng,Shape_Area)) # dump extra stuff

# put on the template grid
mukeys.spdf <- as(mukeys, 'Spatial')
mukeys.spdf <- spTransform(mukeys.spdf, CRS(aeaProj))

soilGrid <- rasterize(mukeys.spdf, template, field='stateMukey',background=NA)

# tidy up the edges (soil polys extend beyond boundary)
bound <- readOGR(gisDir, boundName,verbose=FALSE)
proj4string(bound) <- proj4string(soilGrid)  # already same proj, just needs datum
soilGridMask <- mask(soilGrid, bound)

spplot(soilGridMask)

# export for downstream use
#writeRaster(soilGridMask, paste0(rootDir,'/',soilDir,'/ssurgo_gmd4p_aea30_mask.tif'))
```

455 unique soil types (probably a bit more once split by state)

## Weather
Similarly, sample the weather polygons to the template raster grid

```{r gridWeather}
# load weather
weatherPoly <- readOGR(paste0(rootDir,'/',weatherDir), weatherPolyName, verbose = F)
weatherPoly <- spTransform(weatherPoly, CRS(aeaProj))

# unique gridcells
length(unique(weatherPoly$gridCll))

# rasterize
weatherGrid <- rasterize(weatherPoly, template, field='gridCll',background=NA)
spplot(weatherGrid)

#writeRaster(weatherGrid, paste0(rootDir,'/',weatherDir,'/NLDAS_gmd4p_aea30.tif'))
```

137 unique weather grid cells from NLDAS

## Static Soil & Weather type raster
Combine soil and weather along with padded zeros for easy addition of crop/irrigation status in downstream step.

```{r staticGrid}
# add 8 zeros to the soil grid
static0 <- soilGridMask * 1e8

# add 4 zeros to the weather grid
weatherPadded <- weatherGrid * 1e4

# static master grid!
static <- static0 + weatherPadded

spplot(static, main = 'static soil/weather key')

# write out static grid
# static grid export names
comboDir <- 'combinationGrids'
comboStaticName <- 'StaticVars_gmd4p_aea30.tif'
#writeRaster(static, paste(rootDir, comboDir, comboStaticName, sep='/'))
```

# Dynamic SALUS Combo Maps.
Irrigation status and Crop type change annually. Combine annual CDL/irrigation status with the static soil/weather raster.

Update 11/28/17 - dramatic change to work in vector string space right of the bat. Numeric addition of very large numbers and conversion to matrix to vector was causing problems.

In making the static grid, I had added 4 trailing zeros as placeholders in order to add the cdl-irrigation numbers to the final 4 digits. I also chop that off here.


```{r combos, eval=FALSE}
options(scipen=999) # turn off scientific notation
# load static variable base raster
static <- raster(paste(rootDir, comboDir, comboStaticName, sep='/'))

# convert to vector, then to padded string
staticV <- getValues(static)
staticV.ch0 <- str_pad(staticV, 17, pad = '0')

# remove the 4 trailing zeros no longer needed
staticV.ch <- str_sub(staticV.ch0, start = 1, end = -5)

# output df for combos
annualCombos0 <- setNames(data.frame(matrix(ncol = length(yearsWanted), 
                                           nrow=ncell(static))),
                         c(paste0('x',yearsWanted)))
# add a cell grid number
annualCombos <- cbind(data.frame(gridcell = 1:ncell(static)),
                      annualCombos0)

for(year in yearsWanted){
  # cdl and irrigation annual files
  cdlName <- paste0(rootDir, '/', cdlDir, '/', year, cdlNameSuffix)  
  irrName <- paste0(rootDir, '/', irrDir, '/', year, irrNameSuffix)
  
  # load
  irrigation <- raster(irrName)
  cdl <- raster(cdlName)
  
  # fix the cdl mask exported from GEE (mask 0 values)
  cdl[cdl < 1] <- NA 

  # combine with irrigation and eliminate impossibilities
  cdlpad <- cdl*10
  cdlIrr <- cdlpad + irrigation
  
  # reset any irrigated grassland to dryland
  # reset any irrigated fallow to dryland
  cdlIrr[cdlIrr == 1761] <- 1760
  cdlIrr[cdlIrr == 611] <- 610
  
  # transform to a character vector, padded to 4 digits
  cdlIrrV <- getValues(cdlIrr)
  cdlIrrV.ch <- str_pad(cdlIrrV, 4, pad = '0')
  
  # combine with static string and reset NA's
  combostring <- paste0(staticV.ch,cdlIrrV.ch)
  combostring[str_detect(combostring, 'NA')] <- NA
  
  # # value check - did weird irrigation values get introduced here? NO!
  # combo4 <- str_sub(combostring, start = -4, end = -1)
  # unique(combo4) # 1,4,5,24,61,176 are valid cdl croptypes. now 36 too (afl)
  
  # export single year files
  colname <- paste0('x',year)
  outdata <- structure(data.frame(combostring),names=paste0('x',year))
  save(outdata, file=paste0(rootDir, '/', comboDir, '/', year, comboDynamicName))
  
  # add to master file
  annualCombos[,colname] <- combostring
}

# save combo master
save(annualCombos, file = paste0(rootDir, '/', comboDir, '/', comboMasterFile))
```

# Get Unique Experiments
Get unique combinations. My first idea was to get unique combinations in all years, and run SALUS annually. It turns out the variability in soil water content makes this idea not great. 

update 12/1/2017: no longer relevant (not doing annual runs) and has not been run
with combinatoins that conver irrigated grassland and fallow to dryland

```{r unique, eval=FALSE}
# load
load(file = paste0(rootDir, '/', comboDir, '/', comboMasterFile))

# get all unique combos
uniques <- apply(annualCombos[,-1], 2, function(x) unique(x))

# how many unique combinations each year?
str(uniques)

# get total unique across all years
combo.long <- gather(annualCombos, key = year, value = combo, x2006:x2016) 
allUniques <- unique(combo.long$combo)
length(allUniques)

# save both of these
write.csv(allUniques, row.names = F, file = paste0(rootDir, '/', comboDir, '/', allUnique))
save(uniques, file = paste0(rootDir, '/', comboDir, '/', annualUnique))
```

The new processing form has 24,212 unique combinations across all years.
