==> code_raw_knitted/02.21_SALUS_GetExperimentsContinuous.Rmd <==
---
title: "Unique Combination Trajectories"
author: "Jill Deines"
date: "November 28, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Create SALUS experiment combinations THROUGH TIME: unique soil-climate-crop-irrigation combos

Update 12/6/2017 - Re-run the LEMA to include alfalfa (top7 crops)

Note 11/29/2017 - The county combinations need to be re-done to include grid cell id keys and Experiment ID's still


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/02.21_salus_expts/',
                      cache = TRUE, cache.path='cache/02.21_salus_expts/')
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
library(rgdal)
library(raster)
library(tidyverse)
library(stringr)
```

# Within LEMA
Prototype methods on a subset of raster cells, prioritizing LEMA

## Make a LEMA grid cell key
Get raster template grid cell ids for LEMA in order to select cells from the master data files

```{r lemaKey, eval=FALSE}
# gmd4+ raster grid (based on CDL clipped to study region boundary)
aeaProj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
template <- raster(nrow = 4359, ncol = 7108, crs = aeaProj,
                    xmn = -522870, xmx = -309630, ymn = 1777860, 
                    ymx = 1908630) 
template[] <- 1:ncell(template)

# load lema
gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries'
lema <- readOGR(gisDir, 'Sheridan6_fromKGS_sd6', verbose = FALSE)

# reproject
lema.aea <- spTransform(lema, CRS(aeaProj))
lema.aea$lema <- 1

# rasterize the lema boundary using template
lemaRas <- rasterize(lema.aea, template, field = 'lema', background=NA)

# demonstrate
plot(template, col=colorRampPalette(c('black','blue'))(9))
plot(lemaRas, add=TRUE)

# use the lema raster to mask the template
lemaGridCells <- mask(template, lemaRas)
plot(lemaGridCells)

# export as raster and as vector (single column df)
filterDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids/spatialFilters'
# writeRaster(lemaGridCells, paste0(filterDir,'/gmd4p_template_Sheridan6Only.tif'))

lemaVector <- getValues(lemaGridCells)
length(lemaVector)
sum(!is.na(lemaVector))

lemaVectorOut <- data.frame(lemaCells = lemaVector)
# write.csv(lemaVectorOut, row.names = FALSE,
#           file = paste0(filterDir,'/gmd4p_template_Sheridan_cellKey.csv'))
```

There are a total of 283620 grid cells in the Sheridan 6 Lema.

## Get SALUS Experiments for LEMA
Find unique combinations of soil type, climate grid, and crop+irrigation rotation sequence within LEMA

```{r unique2, eval=FALSE}
# reload lema key
filterDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids/spatialFilters'
lemaKey <- read.csv(paste0(filterDir,'/gmd4p_template_Sheridan_cellKey.csv'))

# load 'annualCombos' data frame
comboDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids'
comboDynamicName <- '_gmd4p_combos_top7.RData'
comboMasterFile <- 'gmd4p_combos_top7_master.RData'
load(file = paste0(comboDir, '/', comboMasterFile))

# keep only the LEMA cells
lemaCombo <- lemaKey %>%
  rename(gridcell = lemaCells) %>%
  filter(!is.na(gridcell)) %>%
  left_join(annualCombos)

# fill in single year missing values with...alfalfa? -------
# filter out years with more than 1 NA
lemaCombo$naCount <- apply(lemaCombo, 1, function(x) sum(is.na(x)))
lemaCombo2 <- lemaCombo[lemaCombo$naCount < 2,] 

# get the weather/soil code for each cell
lemaCombo2$static <- apply(lemaCombo2, 1, function(x) { 
  static <- substr(x[9], start = 1, stop = 13)
  for (m in 1:length(static)){
    if (is.na(static[m])){
      static[m] <- substr(x[11], start = 1, stop = 13)
    }
  }
  return(static)
})

# add alfalfa
ind <- which(is.na(lemaCombo2), arr.ind=TRUE)
lemaCombo2[ind] <- paste0(lemaCombo2$static,'0360')[ind[,1]]

# how many
nrow(lemaCombo2) # there are 267650 "active" cells in Sheridan6

# concatentate all annual columns to get one massive management string
lema_raw <- tidyr::unite(lemaCombo2, 'rawRotations', -gridcell, remove=TRUE)

# unique rotations
lemaUnique <- unique(lema_raw$rawRotations)
length(lemaUnique) # 71401 unique rotation sequences

# assign ExpIDs to unique combinations
ExpIDs <- paste0('L-',1:length(lemaUnique))
lemaExps <- data.frame(ExpID = ExpIDs,
                       ExpCode = lemaUnique)

# and make the ExpID to gridcell link
gridExps <- lemaExps %>%
  left_join(lema_raw, by = c('ExpCode' = 'rawRotations')) %>%
  select(c(gridcell,ExpID))

# write out Key
write.csv(gridExps, row.names=FALSE,
          file = paste0(filterDir, '/uniqueExperiments_gmd4p_Sheridan6_gridcellKey_top7.csv'))
# save key and unique combinations as RData
save(lemaExps,gridExps, 
     file = paste0(filterDir,'/uniqueExperiments_gmd4p_Sheridan6_top7.RData'))
```

That produces 59,346 unique experiments for the LEMA region. It becomes 71401 when you add alfalfa and fille in single NA cells, with alfalfa

## Visualize the experiment coverage
or, what cells have values for the whole study period?

```{r visExptCells}
# gmd4+ template raster grid (based on CDL clipped to study region boundary)
aeaProj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
template <- raster(nrow = 4359, ncol = 7108, crs = aeaProj,
                    xmn = -522870, xmx = -309630, ymn = 1777860, 
                    ymx = 1908630) 
template[] <- 1:ncell(template)
templateVector <- getValues(template)

# aoi
gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries'
aoi <- readOGR(gisDir, 'Sheridan6_fromKGS_sd6', verbose = FALSE)
aoi <- spTransform(aoi, CRS(aeaProj))

# formal aoi cells - use numbers and shorten value column
cells <- lemaCombo2[,1:2]
cells$gridcell <- as.numeric(cells$gridcell)
cells$x2006 <- as.numeric(substr(cells$x2006, start = 1, stop = 3))

# join results to template vector
template.df <- data.frame(gridcell = templateVector)

cellVectors <-  template.df %>%
  left_join(cells, by = 'gridcell')

# rasterize data vector
outRas <- template
outRas[] <- cellVectors[,'x2006']

outRasClip <- crop(outRas, aoi)
spplot(outRasClip, add=T, col.regions='blue', maxPixels = 10000000,
       main = 'Active Cells')
```


# Within 2 Counties surrounding LEMA
Apply to two counties around LEMA to see how many experiments that would be

## Make the grid cell key
Get raster template grid cell ids for LEMA in order to select cells from the master data files

```{r countiesKey}
# gmd4+ raster grid (based on CDL clipped to study region boundary)
aeaProj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
template <- raster(nrow = 4359, ncol = 7108, crs = aeaProj,
                    xmn = -522870, xmx = -309630, ymn = 1777860, 
                    ymx = 1908630) 
template[] <- 1:ncell(template)

# load boundary
gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries'
boundaries0 <- readOGR(gisDir, 'Counties_RRCAextent_Clip', verbose = FALSE)

# subset for lema counties
boundaries <- boundaries0[boundaries0$NAME %in% c('Sheridan','Thomas'),]
plot(boundaries)

# reproject
boundaries.aea <- spTransform(boundaries, CRS(aeaProj))

# rasterize the lema boundary using template
boundRas <- rasterize(boundaries.aea, template, field = 'COUNTYFP', background=NA)

# demonstrate
plot(template, col=colorRampPalette(c('black','blue'))(9))
plot(boundRas, add=TRUE)

# use the county raster to mask the template
countyGridCells <- mask(template, boundRas)
plot(countyGridCells)

# export as raster and as vector (single column df)
filterDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids/spatialFilters'
# writeRaster(countyGridCells, paste0(filterDir,'/gmd4p_template_2counties_sheridanThomas.tif'))

boundVector <- getValues(countyGridCells)
length(boundVector)
sum(!is.na(boundVector))

boundVectorOut <- data.frame(county2Cells = boundVector)
# write.csv(boundVectorOut, row.names = FALSE,
#           file = paste0(filterDir,'/gmd4p_template_2counties_cellKey.csv'))
```

There are a total of 5672072 grid cells in the 2 counties overlapping LEMA (Sheridan and Thomas counties)

## Get SALUS Experiments for 2 counties
Find unique combinations of soil type, climate grid, and crop+irrigation rotation sequence within Sheridan and Thomas counties

```{r unique2counties, eval=FALSE}
# reload county key
filterDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids/spatialFilters'
gridKey <- read.csv(paste0(filterDir,'/gmd4p_template_2counties_cellKey.csv'))

# load 'annualCombos' data frame
comboDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids'
comboDynamicName <- '_gmd4p_combos_top6.RData'
comboMasterFile <- 'gmd4p_combos_top6_master.RData'
load(file = paste0(comboDir, '/', comboMasterFile))

# keep only the LEMA cells
boundCombo <- gridKey %>%
  rename(gridcell = county2Cells) %>%
  filter(!is.na(gridcell)) %>%
  left_join(annualCombos)

# remove rows with any NA's (for example, urban, roads, or other crop classes)
boundCombo2 <- boundCombo[complete.cases(boundCombo),]
nrow(boundCombo2) # there are 4865838 "active" cells in Sheridan and Thomas counties

# concatentate all annual columns to get one massive management string
bound_raw <- tidyr::unite(boundCombo2, 'rawRotations', -gridcell, remove=TRUE)

# unique rotations
boundUnique <- unique(bound_raw$rawRotations)
length(boundUnique) # 59346 unique rotation sequences

# write out unique
save(boundUnique, file = paste0(filterDir,'/uniqueExperiments_gmd4p_2counties.RData'))
```

That produces 815,251 unique experiments for the Sheridan/Thomas regions.



gmd4.ll <- readOGR(paste0(gisDir,'/KS_GMD4.kml'), verbose=FALSE)
