==> code_raw_knitted/02.28_SALUS_nassManagement.Rmd <==
---
title: "Management Parameters"
author: "Jill Deines"
date: "November 14, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Retrieve crop management parameters from the NASS Quickstats API. 

Functions to package up:

* get planting, harvest, maturity.. progress/dates from NASS

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/02.28_salus_management/',
                      cache = TRUE, cache.path='cache/02.28_salus_management/')
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
# for API interaction
library(httr) 
library(jsonlite)

# for extracting median planting dates
library(tidyverse)
library(lubridate)
```

**Directories**

```{r dirVars}
# place to store output
managementDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/management'
```


# Functions: NASS Quick Stats API 

* API website: http://quickstats.nass.usda.gov/api

## Function: Get Progress or densities By Commodity, Year
So far this works for progress dates (ie, planting, maturity, harvest) and row width/population numbers. So I should probably change its name.

Arguments:
Everything currently hard-coded except commodity, category, shortDesc, year, and state 

Current hard coded: 

* program: 'SURVEY' (other option: 'CENSUS'; survey is hardcoded)
* group: 'FIELD CROPS' 
* Domain: 'TOTAL'
* Geographic Level: 'STATE'

Function parameters:

* apiKey: Obtained from http://quickstats.nass.usda.gov/api
* sector: usually 'CROPS'; 'ENVIRONMENTAL' for fertilizer
* Commodity: 'CORN' ('WHEAT', 'SORGHUM','SOYBEANS')
* Category: 'PROGRESS' ('ROW WIDTH', etc)
* Data Item/ short desc: ' - PROGRESS, MEASURED IN PCT PLANTED' ...etc, pasted to crop type but not exactly
* State: 'KANSAS' 
* Year: '2006' = start  year; retrieves data for all subsequent years

Requires: `httr` and `jsonlite`

```{r fun_getPlantingDate}
## function testing variables
# apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
# commodity <- 'CORN'
# shortDesc <- 'CORN - PROGRESS, MEASURED IN PCT PLANTED'
# category <- 'PROGRESS'
# state <- 'KANSAS'
# year <- 2006

getPlantingDate <- function(apikey, sector, commodity, category, shortDesc, state, year){
  # build URL query
  baseurl <- 'http://quickstats.nass.usda.gov/api/api_GET'
  format = 'JSON'
  GETrequest <- paste0(baseurl,
                     '/?key=',apiKey,
                     '&source_desc=', 'SURVEY',
                     '&sector_desc=', sector,
                     '&group_desc=', 'FIELD CROPS',
                     '&commodity_desc=', commodity,
                     '&statisticcat_desc=', category,
                     '&short_desc=', shortDesc,
                     '&domain_desc=', 'TOTAL',
                     '&agg_level_desc=', 'STATE',
                     '&state_name=', state,
                     '&year__GE=', year,
                     '&format=', format)
  
  # if present, replace commas and spaces in url with encodings
  if(grepl(" ", GETrequest))  GETrequest <- gsub(" ", "%20", GETrequest)
  if(grepl(",", GETrequest))  GETrequest <- gsub(",", "%2C", GETrequest)
  
  # retrive data
  req <- GET(GETrequest)
  # check status and warn for errors: 200 means successful, 400 not
  warn_for_status(req, paste('download', state,shortDesc))
  # extract content
  json <- content(req, as = 'text', encoding = 'UTF-8')
  # convert from JSON and extract df from list object
  outputdf <- fromJSON(json)[[1]]
}
```

# Get data

## Planting, Maturity, Harvest Dates
Different date variables (planting, maturity, harvest) are specified in the `cropsWanted` dataframe created below. It's a bit weird but kinda grew organically out of control. Whatevs, it works. Maybe someday the user can just specify a string of short descriptions to get, but they do vary by crop type a bit so eh. 

Short descriptions retrieved from the quickstats.nass.usda.gov manual interface

```{r getDates, eval=FALSE}
# planting date arguments
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
sector <- 'CROPS'
category <- 'PROGRESS'
state <- 'KANSAS'
year <- 2006

# crops wanted and their short description for planting/harvest percentages
dataItemBase_planting <- ' - PROGRESS, MEASURED IN PCT PLANTED'
dataItemBase_harvest <- ' - PROGRESS, MEASURED IN PCT HARVESTED'
dataItemBase_maturity <- ' - PROGRESS, MEASURED IN PCT MATURE'

cropsWanted0 <- data.frame(crop = rep(c('CORN','WHEAT','SORGHUM','SOYBEANS'),3),
                          short = c(paste0('CORN', dataItemBase_planting),
                                    paste0('WHEAT, WINTER', dataItemBase_planting),
                                    paste0('SORGHUM', dataItemBase_planting),
                                    paste0('SOYBEANS', dataItemBase_planting),
                                    paste0('CORN, GRAIN', dataItemBase_harvest),
                                    paste0('WHEAT, WINTER', dataItemBase_harvest),
                                    paste0('SORGHUM, GRAIN', dataItemBase_harvest),
                                    paste0('SOYBEANS', dataItemBase_harvest),
                                    paste0('CORN', dataItemBase_maturity),
                                    paste0('WHEAT, WINTER', dataItemBase_maturity),
                                    paste0('SORGHUM', dataItemBase_maturity),
                                    paste0('SOYBEANS', dataItemBase_maturity)
                                    ))
# but they don't have maturity data for soybeans in KS, so remove
# note wheat only has later years
cropsWanted <- cropsWanted0[-12,]


# stick all crops into a list of crop dfs
dates.list <- list()

for(i in 1:nrow(cropsWanted)){
  dates.list[[i]] <- getPlantingDate(apiKey, sector, cropsWanted[i,'crop'], category,
                                     cropsWanted[i,'short'], state, year)
}

# convert list of data frames to 1 giant dataframe
dates.df <- do.call("rbind",dates.list)

# write out raw NASS data in case the API breaks in the future
write.csv(dates.df, row.names=FALSE, 
  file = paste0(managementDir,'/NASS_dates_planting_maturity_harvest_KS_2006-2017_4crops_RAW.csv'))
```

Extract date of 50% planting/maturity/harvest for each crop and year. After grouping by crop and year, sort each group by the absolute distance of the % planted from 50, and take the top record.

Do some data munging to make a nice df.

```{r niceDates}
# load date data downloaded from NASS
dates.raw <- read.csv(paste0(managementDir,'/NASS_dates_planting_maturity_harvest_KS_2006-2017_4crops_RAW.csv'))

# get the "median" planting, maturity, and harvest dates in a nicely formatted table
dates.50 <- dates.raw %>% 
  select(c(state_name, year, week_ending, short_desc, commodity_desc, Value)) %>% 
  group_by(short_desc, year) %>%
  arrange(abs(Value - 50)) %>%         # calculate distance from 50% and sort on this
  slice(1) %>%                         # take the top record (closest to 50%)
  # and format into nice look up table:
  mutate(dateType = gsub("^.* ", "", short_desc)) %>%
  ungroup() %>%
  select(-c(short_desc, Value)) %>%     # remove columns to allow row collapse
  spread(dateType, week_ending) %>%
  rename(plantingDate_50 = PLANTED,
         harvestDate_50 = HARVESTED,
         maturityDate_50 = MATURE,
         crop = commodity_desc) %>%
  mutate(plantingDOY_50 = yday(ymd(plantingDate_50)),
         harvestedDOY_50 = yday(ymd(harvestDate_50)),
         matureDOY_50 = yday(ymd(maturityDate_50))) %>%
  arrange(year,crop)

# # export
# write.csv(dates.50, row.names=FALSE, 
#   file = paste0(managementDir,'/NASS_dates_planting_maturity_harvest_KS_2006-2017_4crops_MEDIANs.csv'))

# make a plot
dates.50.long <- dates.50 %>% select(-c(harvestDate_50,maturityDate_50,plantingDate_50)) %>%
  gather(key = dateType, value = DOY, plantingDOY_50:matureDOY_50)

ggplot(dates.50.long, aes(x = year, y = DOY, color = dateType)) +
  geom_point() + geom_line() +
  facet_wrap(~crop) + theme_bw() +
  ggtitle('Dates of 50% planted, maturity, and harvest, KS')
```

## Row Width, Density, Fertilizer

### Row width
Not available for sorghum or wheat

```{r getRowWidth, eval=FALSE}
# planting date arguments
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
sector <- 'CROPS'
category <- 'ROW WIDTH'
state <- 'KANSAS'
year <- 2006

# crops wanted and their short description for planting/harvest percentages
dataItemBase_row <- ' - ROW WIDTH, MEASURED IN INCHES / ROW'

cropsWanted2 <- data.frame(crop = c('CORN','SOYBEANS'),
                          short = c(paste0('CORN, GRAIN', dataItemBase_row),
                                    paste0('SOYBEANS', dataItemBase_row)))


# stick all crops into a list of crop dfs
width.list <- list()

for(i in 1:nrow(cropsWanted)){
  width.list[[i]] <- getPlantingDate(apiKey, sector, cropsWanted2[i,'crop'], category,
                                     cropsWanted2[i,'short'], state, year)
}

# convert list of data frames to 1 giant dataframe
width.df <- do.call("rbind",width.list)

# write out raw NASS data in case the API breaks in the future
write.csv(width.df, row.names=FALSE, 
  file = paste0(managementDir,'/NASS_rowWidth_KS_2012-2017_cornSoy_RAW.csv'))
```

### Plant densities
These start to get spotty and weird:

* Plant density is currently only available for corn. It's only available for irrigated/non-irrigated corn separately in Nebraska.

```{r getDensities, eval=FALSE}
# generic arguments for this project
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
state <- 'KANSAS'
year <- 2006

# density arguments
sector <- 'CROPS'
category <- 'PLANT POPULATION'
crop <- 'CORN'
short <- 'CORN, GRAIN - PLANT POPULATION, MEASURED IN PLANTS / ACRE'

# only a single df to retrieve
density.df <- getPlantingDate(apiKey, sector, crop, category, short, state, year)

# write out raw NASS data in case the API breaks in the future
write.csv(density.df, row.names=FALSE, 
  file = paste0(managementDir,'/NASS_plantDensity_KS_2006-2017_corn_RAW.csv'))
```

### Series of Ear/Head/Pod Counts
There's also ear (corn), head (wheat), and pod counts (soy). There doesn't seem to be any related information for sorghum.

```{r getCounts, eval=FALSE}
# planting date arguments
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
sector <- 'CROPS'
category <- 'PROGRESS'
state <- 'KANSAS'
year <- 2006

# CROP, CATEGORY, AND SHORT/DATA ITEM VARIABLES
cropsWanted3 <- data.frame(crop = c('CORN','CORN','WHEAT','SOYBEANS'),
                           category = c('EAR COUNT','EAR COUNT', 'HEAD COUNT', 'POD COUNT'),
                           short = c('CORN, GRAIN - EAR COUNT, MEASURED IN EARS / ACRE',
                                     'CORN, GRAIN, IRRIGATED - EAR COUNT, MEASURED IN EARS / ACRE',
                                     'WHEAT, WINTER - HEAD COUNT, MEASURED IN HEAD / SQ FT',
                                     'SOYBEANS - POD COUNT, MEASURED IN PODS / 18 SQ FT'))



# stick all crops into a list of crop dfs
count.list <- list()

for(i in 1:nrow(cropsWanted3)){
  count.list[[i]] <- getPlantingDate(apiKey, sector, cropsWanted3[i,'crop'], 
                                     cropsWanted3[i,'category'],
                                     cropsWanted3[i,'short'], state, year)
}

# convert list of data frames to 1 giant dataframe
count.df <- do.call("rbind",count.list)

# write out raw NASS data in case the API breaks in the future
write.csv(count.df, row.names=FALSE, 
  file = paste0(managementDir,'/NASS_countsFruitingBodies_KS_2006-2017_3crops_RAW.csv'))
```

### Fertilizer
It's starting to feel silly that I didn't retrieve all of these variables at once...

only for corn; also hard to tell what info i need. tabling for now.

```{r getFertilizer, eval=FALSE}
# generic arguments for this project
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
state <- 'KANSAS'
year <- 2006

# density arguments
sector <- 'ENVIRONMENTAL'
category <- 'APPLICATIONS'
crop <- 'CORN'
```

# Synthesize a Master Management Look-up Table
put 'em together

Right now, just:

* row width
* corn density
* random counts per area (fruiting bodies)

```{r masterManagement}
# load output from chunks above
dates <- read.csv(paste0(managementDir,'/NASS_dates_planting_maturity_harvest_KS_2006-2017_4crops_MEDIANs.csv'))

fruitingBodies <- read.csv(paste0(managementDir,'/NASS_countsFruitingBodies_KS_2006-2017_3crops_RAW.csv'))
density <- read.csv(paste0(managementDir,'/NASS_plantDensity_KS_2006-2017_corn_RAW.csv'))
rowWidth <- read.csv(paste0(managementDir,'/NASS_rowWidth_KS_2012-2017_cornSoy_RAW.csv'))

# clean row/density etc
spacing0 <- rbind(rowWidth, rbind(fruitingBodies, density))

spacing <- spacing0 %>%
  # make nicer columns for county/state fips and irrigated/not/all statuses
  mutate(type = sub(" .*", "", prodn_practice_desc)) %>%
  rename(crop = commodity_desc, 
         variable = statisticcat_desc,
         density = ) %>%
  # drop most columns
  select(c(state_name, year, crop, type, variable, short_desc, 
           Value, unit_desc)) %>%
  arrange(crop, Value, year) 
  


```
