==> code_raw_knitted/02.30_SALUS_nassYields.Rmd <==
---
title: "NASS County Yields"
author: "Jill Deines"
date: "November 15, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Retrieve crop yields from the NASS Quickstats API:

* annual survey by county
* annual survey by state, since a lot of counties dont have survey data
* Census data by county


Functions to package up:

* get yields from NASS - incorporated with the management function, or a separate one (many small, one large?)

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/02.30_salus_nassYields/',
                      cache = TRUE, cache.path='cache/02.30_salus_nassYields/')
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
# for API interaction
library(httr) 
library(jsonlite)

# to pull county names
library(sf)

library(tidyverse)
```

**Directories**

```{r dirVars}
# place to store output
yieldDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/yield_nass'

# studay area boundary with counties
boundFile <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries/GMD4_plus_byCounty.shp'
```


# Functions: NASS Quick Stats API 

* API website: http://quickstats.nass.usda.gov/api

## Function: Get Yield by Commodity, County, Year
part of the set of functions i've made for NASS that really should become 1 or 2.

Current hard coded: 

* sector: 'CROPS'
* group: 'FIELD CROPS' 
* Domain: 'TOTAL'
* Geographic Level: 'STATE'

Function parameters:

* apiKey: Obtained from http://quickstats.nass.usda.gov/api
* program: 'SURVEY' (other option: 'CENSUS')
* Commodity: 'CORN' ('WHEAT', 'SORGHUM','SOYBEANS')
* Category: 'YIELD' 
* Data Item/ short desc: ' - PROGRESS, MEASURED IN PCT PLANTED' ...etc, pasted to crop type but not exactly
* State: 'KS' - 2 letter code 
* county: county 3-digit ANSI code
* Year: '2006' = start  year; retrieves data for all subsequent years

Requires: `httr` and `jsonlite`

```{r fun_getYield}
# # function testing variables
# apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
# program <- 'SURVEY'
# commodity <- 'CORN'
# shortDesc <- 'CORN, GRAIN, IRRIGATED - YIELD, MEASURED IN BU / ACRE'
# state <- 'KS'
# county <- 179
# year <- 2006

getCountyYieldNASS <- function(apikey, program, commodity, shortDesc, state, 
                               county, year){
  # build URL query
  baseurl <- 'http://quickstats.nass.usda.gov/api/api_GET'
  format = 'JSON'
  GETrequest <- paste0(baseurl,
                     '/?key=',apiKey,
                     '&source_desc=', program,
                     '&sector_desc=', 'CROPS',
                     '&group_desc=', 'FIELD CROPS',
                     '&commodity_desc=', commodity,
                     '&statisticcat_desc=', 'YIELD',
                     '&short_desc=', shortDesc,
                     '&domain_desc=', 'TOTAL',
                     '&agg_level_desc=', 'COUNTY',
                     '&state_alpha=', state,
                     '&county_ansi=', county,
                     '&year__GE=', year,
                     '&format=', format)
  
  # if present, replace commas and spaces in url with encodings
  if(grepl(" ", GETrequest))  GETrequest <- gsub(" ", "%20", GETrequest)
  if(grepl(",", GETrequest))  GETrequest <- gsub(",", "%2C", GETrequest)
  
  # retrive data
  req <- GET(GETrequest)
  # check status and throw stop/error: 200 means successful
  warn_for_status(req, paste('download', state,county,shortDesc))
  # extract content
  json <- content(req, as = 'text', encoding = 'UTF-8')
  # convert from JSON and extract df from list object
  outputdf <- fromJSON(json)[[1]]
}
```

## function: get yield by state
copying this function is getting out of hand.

```{r fun_getStateYield}
# function testing variables
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
program <- 'SURVEY'
commodity <- 'CORN'
shortDesc <- 'CORN, GRAIN, IRRIGATED - YIELD, MEASURED IN BU / ACRE'
state <- 'KS'
year <- 2006

getStateYieldNASS <- function(apikey, program, commodity, shortDesc, state, 
                               year){
  # build URL query
  baseurl <- 'http://quickstats.nass.usda.gov/api/api_GET'
  format = 'JSON'
  GETrequest <- paste0(baseurl,
                     '/?key=',apiKey,
                     '&source_desc=', program,
                     '&sector_desc=', 'CROPS',
                     '&group_desc=', 'FIELD CROPS',
                     '&commodity_desc=', commodity,
                     '&statisticcat_desc=', 'YIELD',
                     '&short_desc=', shortDesc,
                     '&domain_desc=', 'TOTAL',
                     '&agg_level_desc=', 'STATE',
                     '&state_alpha=', state,
                     '&year__GE=', year,
                     '&format=', format)
  
  # if present, replace commas and spaces in url with encodings
  if(grepl(" ", GETrequest))  GETrequest <- gsub(" ", "%20", GETrequest)
  if(grepl(",", GETrequest))  GETrequest <- gsub(",", "%2C", GETrequest)
  
  # retrive data
  req <- GET(GETrequest)
  # check status and throw stop/error: 200 means successful
  warn_for_status(req, paste('download', state, shortDesc))
  # extract content
  json <- content(req, as = 'text', encoding = 'UTF-8')
  # convert from JSON and extract df from list object
  outputdf <- fromJSON(json)[[1]]
}
```

# Get Yield Data

## Get Counties of Interest
Pull the county FIPS/state from the project boundary shapefile by county

```{r getCounties}
# load 
boundByCounty <- read_sf(boundFile) %>% as.data.frame()

# get state and 3 digit fips
countiesToGet <- boundByCounty %>%
  mutate(fips3 = substr(fips5, 3,5)) %>%
  select(c(abb, fips3))
```

## Query Parameters
Make a query parameter data frame

### short description types
combinations for the data items based on crop type
```{r makeShorts}
# yield data types short description bases: all, irrigated, rainfed
yieldTypes <- paste0(c('', ', IRRIGATED', ', NON-IRRIGATED'), ' - YIELD, MEASURED IN BU / ACRE')

# crop short desc prefixes
shortPrefixes <- c('CORN, GRAIN', 'SORGHUM, GRAIN', 'SOYBEANS','WHEAT')

# full set of short descriptions needed
shortDescs <- sort(paste0(rep(shortPrefixes,3), yieldTypes))

# df of all combos
cropYieldsToGet <- data.frame(crop = c(rep('CORN',3), rep('SORGHUM',3), 
                                       rep('SOYBEANS',3), rep('WHEAT',3)),
                              short = shortDescs)
```

## Retrieve data
behold the nested for lopp

```{r getYield, eval=FALSE}
# parameters
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
program <- 'SURVEY'
year <- 2006

# list to store output
yield.list <- vector('list',nrow(countiesToGet)*nrow(cropYieldsToGet))

# get data
x <- 1  # start list indexing at 1
for (i in 1:nrow(countiesToGet)){
  # set variables
  state <-  countiesToGet[i,'abb']
  county <- countiesToGet[i,'fips3']
  
  for(m in 1:nrow(cropYieldsToGet)){
    crop <- as.character(cropYieldsToGet[m,'crop'])
    short <- as.character(cropYieldsToGet[m,'short'])
    
    # retrieve data!
    yield.list[[x]] <- getCountyYieldNASS(apikey, program, crop, 
                               short, state, county, year)
    x <- x + 1  # advance list index
  }      
}
```

A quick check of the warnings indicates that for those counties, there's no data for that crop/short description for the years 2006+

### Combine and Export Raw Data


```{r formatYield, eval=FALSE}
# remove bad queries from list
hasData <- sapply(yield.list, function(x) class(x)=='data.frame')
yields.list2 <- yield.list[hasData]

# convert list of data frames to 1 giant dataframe
yields.df <- do.call("rbind",yields.list2)

# write out raw NASS data in case the API breaks in the future
write.csv(yields.df, row.names=FALSE, 
  file = paste0(yieldDir,'/NASS_yields_KS_2006-2017_4crops_RAW.csv'))
```


## Make It Nice
Including converting from bushels / acre to kg / hectare. Weights of crops per bushel:

UNC site says 

* corn, shelled: 56
* corn, ear: 70
* corn (in ear):35

crop | pounds/bushel | 1 bu/acre = 
--------------- | ------ | --------
corn/sorghum | 56 | 62.89 kg/ha
soy/wheat | 60 | 67.11 kg/ha

From

* https://www.extension.iastate.edu/agdm/wholefarm/html/c6-80.html
* http://www.unc.edu/~rowlett/units/scales/bushels.html

```{r formatData}
yield.raw <- read_csv(paste0(yieldDir,'/NASS_yields_KS_2006-2017_4crops_RAW.csv'))

# conversions
convertKey = data.frame(crop = c('CORN','SORGHUM', 'SOYBEANS', 'WHEAT'),
                        conversion = c(62.89,62.89,67.11,67.11))

# simplify
yield.clean <- yield.raw %>%
  # make nicer columns for county/state fips and irrigated/not/all statuses
  mutate(type = sub(" .*", "", prodn_practice_desc),
         fips5 = paste0(state_fips_code, county_code)) %>%
  rename(crop = commodity_desc, yield_bu_acre = Value) %>%
  left_join(convertKey) %>%
  mutate(yield_kg_ha = yield_bu_acre * conversion) %>%
  # drop most columns
  select(c(year, fips5, county_name, crop, type, yield_kg_ha, yield_bu_acre, 
           state_name)) %>%
  arrange(year, fips5, crop)

#export
write.csv(yield.clean, row.names=FALSE, 
  file = paste0(yieldDir,'/NASS_yields_KS_2006-2017_4crops_CLEANED.csv'))
```


## Visualize

Overall regional yields by crop type and irrigation status: English units

```{r regionalYields}
# format data
regional_english <- yield.clean %>%
  group_by(year, crop, type) %>%
  do(data.frame(t(quantile(.$yield_bu_acre, probs = c(.1, .25, .5, .75, .9),na.rm=TRUE))))

# plot
ggplot(regional_english, aes(x=year, y = X50., color = type)) +
  geom_ribbon(aes(x = year, ymin = X25., ymax = X75., fill=type), alpha = .4) +
  geom_line() +
  facet_wrap(~crop) +
  ylab('Yield (BU/ACRE)') + xlab('') + 
  ggtitle('Median and IQR Yield, GMD4+ Counties') +
  theme_bw() + theme(legend.position = 'bottom', legend.title = element_blank())
```

Overall regional yields by crop type and irrigation status: metric units

```{r regionalYields_metric, fig.width = 5, fig.height = 4.25}
# format data
regional_metric <- yield.clean %>%
  group_by(year, crop, type) %>%
  do(data.frame(t(quantile(.$yield_kg_ha, probs = c(.1, .25, .5, .75, .9),na.rm=TRUE))))

# plot
ggplot(regional_metric, aes(x=year, y = X50., color = type)) +
  geom_ribbon(aes(x = year, ymin = X25., ymax = X75., fill=type), alpha = .4) +
  geom_line() +
  facet_wrap(~crop) +
  ylab('Yield (kg/ha)') + xlab('') + 
  ggtitle('Median and IQR Yield, GMD4+ Counties') +
  theme_bw() + theme(legend.position = 'bottom', legend.title = element_blank())
```

# State Yield
Get yields at the state level since county data doesn't exist each year.

## Download annual state yields

```{r stateSurvey, eval=FALSE}
# parameters
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
program <- 'SURVEY'
state <- 'KS'
year <- 2006

# list to store output
state.list <- vector('list',nrow(cropYieldsToGet))

# get data
for (i in 1:nrow(cropYieldsToGet)){
  # set variables
  crop <- as.character(cropYieldsToGet[i,'crop'])
  short <- as.character(cropYieldsToGet[i,'short'])
    
  # retrieve data!
  state.list[[i]] <- getStateYieldNASS(apikey, program, crop, 
                               short, state, year)
}

# convert list of data frames to 1  dataframe
state.df <- do.call("rbind",state.list)

# write out raw NASS data in case the API breaks in the future
write.csv(state.df, row.names=FALSE, 
  file = paste0(yieldDir,'/NASS_yields_STATE_KS_2006-2017_4crops_RAW.csv'))
```

## Make It Nice

```{r formatData2}
state.raw <- read_csv(paste0(yieldDir,'/NASS_yields_STATE_KS_2006-2017_4crops_RAW.csv'))

# conversions
convertKey = data.frame(crop = c('CORN','SORGHUM', 'SOYBEANS', 'WHEAT'),
                        conversion = c(62.89,62.89,67.11,67.11))
# simplify
state.clean <- state.raw %>%
  # remove forecast values
  filter(reference_period_desc == 'YEAR') %>%
  # make nicer columns for county/state fips and irrigated/not/all statuses
  mutate(type = sub(" .*", "", prodn_practice_desc)) %>%
  rename(crop = commodity_desc, yield_bu_acre = Value) %>%
  # convert to kg/ha
  left_join(convertKey) %>%
  mutate(yield_kg_ha = yield_bu_acre * conversion) %>%
  # drop most columns
  select(c(year, crop, type, yield_kg_ha, yield_bu_acre, state_name)) %>%
  arrange(year, crop, type)
 

#export
write.csv(state.clean, row.names=FALSE, 
  file = paste0(yieldDir,'/NASS_yields_STATE_KS_2006-2017_4crops_CLEANED.csv'))
```


## Visualize

Overall STATE yields by crop type and irrigation status:

```{r stateYields, fig.width = 5, fig.height = 4.25}
# plot
ggplot(state.clean %>% filter(year < 2017), 
       aes(x=year, y = yield_bu_acre, color = type)) +
  geom_line() +
  facet_wrap(~crop) +
  ylab('Yield (BU/ACRE)') + xlab('') + ggtitle('Kansas State Yields') +
  theme_bw() + theme(legend.position = 'bottom', legend.title = element_blank())

ggplot(state.clean %>% filter(year < 2017), 
       aes(x=year, y = yield_kg_ha, color = type)) +
  geom_line() +
  facet_wrap(~crop) +
  ylab('Yield (kg/ha)') + xlab('') + ggtitle('Kansas State Yields') +
  theme_bw() + theme(legend.position = 'bottom', legend.title = element_blank())
```
