==> code_raw_knitted/02.50_SALUS_writeExperiments_test.Rmd <==
---
title: "Generate Salus Experiment Files"
author: "Jill Deines"
date: "November 23, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Make Experiment Files; here, a continuous run of wheat. Also makes a test batch and shell file for the HPCC (.bat and .sh), and briefly looks at soil water content among years to inform annual vs continuous decision.

Note 12/3/17: salustools functions have been updated, this shouldn't work anymore


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/02.50_salus_expt/',
                      cache = TRUE, cache.path='cache/02.50_salus_expt/')
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
library(tidyverse)
library(stringr)
library(sf)
library(salustools) # from Jill's github
library(lubridate)
library(viridis)
```

**Directories**
meh, there directories set all throughout this script

```{r dirs}
testDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/SALUS/testGUIrun/continuousWheat'
testFile <- 'wheat_continuous_newFuns.xdb.xml'
```

## Extract some management info

```{r getManagement}
# load date file
managementDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/management'
dates <- read.csv(paste0(managementDir,'/NASS_dates_planting_maturity_harvest_KS_2006-2017_4crops_MEDIANs.csv'))

# extract wheat means
wheatDates <- dates %>% filter(crop == 'WHEAT')
summary(wheatDates$plantingDOY_50)
summary(wheatDates$harvestedDOY_50, na.rm = TRUE)
```


# Test Xdb Functions
In the `salustools` package, I made a set of hopefully flexible functions to programmaticlly create a wide range of Experiment files. It was pieced together from examining the example .xdb.xmls', Lydia Rill and Brian Baer's python function, Lin's example, and SALUS help files

I think I will break this up into tiny buildable functions:

* write_xdb_topMatter
* write_xdb_experiment
* write_xdb_rotation (with argument for first rotation: starts section)
  * write_xdb_mPlanting - write planting parameters for crop rotation
  * write_xdb_mFertilize -
  * write_xdb_mHarvest
  * etc
* write_xdb_bottomMatter - closes above components as specified

Functions that could be added?

* write_xdb_measured_data - I don't think I have measured data yet? would be entered below experiment at the same level of Rotation Components

## Test: 2 wheat years
use functions to make an Experiment file for 2 wheat 

### Function parameters
list of parameter arguments. in production, the would be tied together 

```{r test_2wheat_parameters}
outFile <- paste0(testDir,'/', testFile)
# Experiment
ExpID <- 1
RunTitle <- 'wheatTest2'
startYear <- 2006
NYrs <- 3
startDOY <- 270
WthID <- '1001'
WthFile <- '1001.wdb.xml'
SoilID <- 'KS1017570'
SoilFile <- 'KS.sdb.xml'
CropFile <- 'cropsn29Dec2016.cdb.xml'
# rotation components
OrderNum <- 1
Title <- 'Wheat'
irrig <- 'N'  # 'N' - not, 'A' - auto, + more options
fertilize <- 'R' # R = reported date, 'A' - auto, 'N' - not
tillage <- 'N' # N = none, 'A' = automatic,, etc
harvest <- 'R' #R - reported, 'M' = at maturity
# management - planting
cropModel <- 'C' # C = complex, S = simple
species <- 'WH' # 'WH' = wheat
cultivar <- 'P25R40'
year <- 2006
PlantDOY <- 276 # median
# management - fertilizer
year_f <- 2007
fertDOY <- 75 
Nfert <- 78.47 # kg / ha
# management - harvest
year_h <- 2007
harvestDOY <- 178 
```

Build the Experiment Files from my functions!

```{r test_2wheat, eval=FALSE}
write_xdb_topMatter(outFile)
write_xdb_experiment(outFile, ExpID, RunTitle, startYear, NYrs, startDOY, 
                     WthID, WthFile, SoilID, SoilFile, CropFile)
# year 1
write_xdb_rotation(outFile, OrderNum, Title, irrig, fertilize, tillage, harvest)
write_xdb_mPlanting(outFile, cropModel, species, cultivar, year, PlantDOY, 455,17)
write_xdb_mFertilize(outFile, year_f, fertDOY, Nfert)
write_xdb_mHarvest_reported(outFile, year_h, harvestDOY, closeComponent = 'Y')
# year 2
write_xdb_rotation(outFile, OrderNum+1, Title, irrig, fertilize, tillage, harvest)
write_xdb_mPlanting(outFile, cropModel, species, cultivar, year+1, PlantDOY, 455, 17)
write_xdb_mFertilize(outFile, year_f+1, fertDOY, Nfert)
write_xdb_mHarvest_reported(outFile, year_h+1, harvestDOY, closeComponent = 'Y')
# close out
write_xdb_bottomMatter(outFile, closeComponent = 'N', closeRotation = 'Y', 
                        closeExperiment = 'Y', writeVersion = 'Y', 
                        closeXDB = 'Y')
```

Report: It totally worked!

# Run Continuous Wheat

Run 10 years of wheat to see when a good time to re-start the year might be, and at what value soil water should be initialized at.

## Extract real world wheat scenarios
Load my unique combinations in order to select soils which commonly have wheat on them

```{r getCombos}
# dirs
rootDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus'
comboDir <- 'combinationGrids'
allUnique <- 'allUnique_gmd4p_top6_2006-2016.csv'

# load unique combinations
uniques <- read_csv(paste0(rootDir, '/', comboDir, '/', allUnique))
names(uniques) <- 'rasterCode'
uniques <- uniques[-1,]
  
# parse into informative columns and extract dryland wheat in KS
uniqueSpread <- uniques %>%
  mutate(irr = str_sub(rasterCode, start = -1, end = -1),
         cdl = str_sub(rasterCode, start = -4, end = -2),
         nldas = str_sub(rasterCode, start = -8, end = -5),
         mukey = str_sub(rasterCode, start = -15, end = -9),
         state = str_sub(rasterCode, start = -17, end = -16)) %>%
  filter(cdl == '024' & irr == '0' & state == '20')

# keep only NLDAS grids within LEMA ------------------------------------------
# load nldas
weatherDir <- 'weather/nldas_Jill/nldas_centroidExtract'
weatherPolyName <- 'NLDAS_clipped_gmd4p.shp'
nldasPolys <- read_sf(paste(rootDir, weatherDir, weatherPolyName, sep = '/'))
# load lema
lema <- read_sf('C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries/Sheridan6_trace.shp') %>% st_transform(4326)
# get nldas cells
lemaNldas <- st_intersection(nldasPolys,lema)
lemagrids <- as.character(lemaNldas$gridCll)

uniqueLema <- uniqueSpread %>%
  filter(nldas %in% lemagrids)

# that gives 96 combinations. 
# actually, I just identified the largest soil polygon within LEMA in ArcGIS: 
# biggest: 2759231
# next biggest: 1150000

bestSoil <- uniqueLema %>% filter(mukey %in% c('2759231','1150000'))
bestSoil
```

Run these 11 experiments!

## Generate Experiment file for these dominant soils in LEMA
Run all 11 experiments to also see how my `write_xdb_*` functions scale when in use for actual scenarios

```{r lemaWheat, eval=FALSE}
# table of annual specifications: median planting and harvest dates, fertililzation,
# and other component parameters
annualSpecs <- wheatDates %>%
  mutate(plantYear = year(ymd(plantingDate_50)),
         harvYear = year(ymd(harvestDate_50))) %>%
  select(c('year','crop','plantingDOY_50','harvestedDOY_50','plantYear','harvYear')) %>%
  # add a generic fertilization date and amount
  mutate(fertDOY = 75,
         Nfert = 78.47,
         irrig = 'N',
         fertilize = 'R',
         tillage = 'N',
         harvest = 'R',
         cropModel = 'C',
         species = 'WH',
         cultivar = 'P25R40') %>%
  filter(year <= 2016)

# Experiment Table
exp_master <- bestSoil %>%
  mutate(soilId = paste0('KS',mukey)) %>%
  select(c(nldas, soilId)) %>%
  mutate(soilfp = 'KS.sdb.xml',
         weatherfp = paste0(nldas,'.wdb.xml'),
         ExpID = 1:nrow(bestSoil),
         runTitle = paste0('Wheat-',soilId,'-',nldas),
         startYear = 2005,
         Nyrs = 11,
         startDOY = 265,
         cropfp = 'cropsn29Dec2016.cdb.xml')


outFile2 <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/SALUS/testGUIrun/lemaWheatContinuous/lema_wheat_continuous.xdb.xml'

# Write the file! --------------------------------------------------------------

# initialize the file
write_xdb_topMatter(outFile2)

# for each experiment:
for(i in 1:nrow(exp_master)){  
  # write each set of experiment parameters
  write_xdb_experiment(outFile2, exp_master$ExpID[i], exp_master$runTitle[i], 
                       exp_master$startYear[i], exp_master$Nyrs[i], exp_master$startDOY[i], 
                       exp_master$nldas[i], exp_master$weatherfp[i], exp_master$soilId[i], 
                       exp_master$soilfp[i], exp_master$cropfp[i])
  # write rotation and management parameters for each year
  for(m in 1:nrow(annualSpecs)){
    # rotation parameters
    write_xdb_rotation(outFile2, m, annualSpecs$crop[m], annualSpecs$irrig[m], 
                       annualSpecs$fertilize[m], annualSpecs$tillage[m], 
                       annualSpecs$harvest[m])
    # management: planting
    write_xdb_mPlanting(outFile2, annualSpecs$cropModel[m], annualSpecs$species[m], 
                        annualSpecs$cultivar[m], annualSpecs$plantYear[m], 
                        annualSpecs$plantingDOY_50[m])
    # management: fertilize
    write_xdb_mFertilize(outFile2, annualSpecs$harvYear[m], annualSpecs$fertDOY[m],
                         annualSpecs$Nfert[m])
    # management: harvest
    write_xdb_mHarvest(outFile2, annualSpecs$harvYear[m], annualSpecs$harvestedDOY_50[m], 
                       closeComponent = 'Y')
  }  
  # close rotations and experiment
  write_xdb_bottomMatter(outFile2, closeRotation = 'Y', closeExperiment = 'Y')
}  
# close out
write_xdb_bottomMatter(outFile2, writeVersion = 'Y', closeXDB = 'Y')
```

Hey, that worked like a charm, first go!

## Write HPCC batch Files
This is done as a post hoc test on making functions to write the .bat and .sh files needed for HPCC runs. The results in the section below are from running the .xdb manually in the SALUS GUI.

### HPC shell script
only one sh file here, for one experiment

```{r hpccSh}
shDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/SALUS/testHPCCrun'
hpcHomeDir <- '/mnt/home/deinesji/Example_SALUS_wheat/'
hpcOutDir <- '/mnt/home/deinesji/Example_SALUS_wheat/results/'
xdb <- 'lema_wheat_continuous'
sdb <- "KS"
wdbZip <- 'testHPCCrun.tar.gz'
DayVars <- 'ExpID,Title,SpeciesID,GWAD,IRRC,CWAD,DRNC,PREC,LAI'
SeaVars <- 'ExpID,Title,SpeciesID,GWAD,IRRC,CWAD,DRNC,PREC'
walltime <- '01:00:00'
memory <- '2000mb'

salustools::write_HPC_shell(shDir, hpcHomeDir, hpcOutDir, xdb, sdb, wdbZip, DayVars,
                 SeaVars, walltime, memory)
```

### HPC bash bat script
a bit redundant since only 1 sh file, but sends the sh file to queue

```{r hpccbat}
shNames <- 'lema_wheat_continuous.sh'
fileOut <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/SALUS/testHPCCrun/lematest.bat'

# write out .bat file
write_HPC_bat(shNames, fileOut)
```



## Examine results
Specifically interested in the soil water to determine good start dates for annual reinitialization

### Soil water
Salus automatically chops subsurface into 20 layers denoted here: http://salusmodel.glg.msu.edu/tools/layers.html. Note the sw variable units is soil water content, m^3 / m^3

Here, I get the weighted average for the top 1.25 meter (first 9 layers) using:

* SW_tot = SW_lay * thick_lay/sum(thick_lay)

```{r salusOutput}
# load DAILY output
resultsDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/SALUS/testGUIrun/lemaWheatContinuous/results'
daily <- read_csv(paste0(resultsDir, '/salus.day.csv')) %>%
  select(c('Title','Year','DOY','PREC','SW(1)','SW(2)','SW(3)','SW(4)','SW(5)',
           'SW(6)','SW(7)','SW(8)','SW(9)'))

# set layer thicknesses, in meters
layerKey <- data.frame(layerNum = as.character(c(1:9)),
                       layerThick = c(2,5,8,11,14,17,20,23,25)/100)
sumThick <- sum(layerKey$layerThick)

# longify data, extract layer number as a column, join with layer thickness
sw_long <- daily %>% gather(key = swLayer, value = swContent.m3.m3, `SW(1)`:`SW(9)`) %>%
  mutate(layerNum = str_sub(swLayer, start = -2, end = -2)) %>%
  left_join(layerKey)

# get the weighted average for each DOY
sw_total <- sw_long %>%
  mutate(swTot_lay = swContent.m3.m3 * (layerThick / sumThick)) %>%
  group_by(Title, Year, DOY) %>%
  summarise(SWTot = sum(swTot_lay))
```

#### Plot: soil water

```{r plotSW_all}
ggplot(sw_total, aes(x = DOY, y = SWTot, color = as.factor(Year))) +
  geom_line() +
  scale_color_viridis(direction = -1, discrete=TRUE) +
  facet_wrap(~Title)

# just one Experiment
one <- sw_total %>% filter(Title == 'Wheat-KS2759231-1124')
ggplot(one, 
       aes(x = DOY, y = SWTot, color = as.factor(Year))) +
  geom_line() +
  scale_color_viridis(direction = -1, discrete=TRUE) +
  ggtitle('Surface water content by Year, dryland wheat')
```

Well, that's enough of an answer - soil water is highly variable among years, I should shoot for continuous.
