==> code_raw_knitted/03.20_SALUS_Results_Process.Rmd <==
---
title: "Process Salus Results"
author: "Jill Deines"
date: "December 4, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Build SALUS results workflow to process LEMA output back into a spatial grid

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/03.20_7_salus_results_process/',
                      cache = FALSE)
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
# trying to include package::function but the order of these two packages 
# will matter if I forget
library(raster)  
library(tidyverse)
library(rgdal)
#library(devtools)
#install_github("jdeines/salustools")
library(salustools)
```

**Directories and User Variables**

```{r dirs}
# directory for cleaned/summarized to annual values salus output
resultsDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/experiments/7_LEMA_hist_realEst_v02/results'
# subfolder for maps output
subf <- 'maps'

# run name
runname <- '7_lema_historic_realEst_v02'

# time frame
startYear <- 2006 # excluding wheat planting, aka first year of harvest
endYear <- 2016

# load unique experiments and grid cell to Experiment code key
#comboDir <- 'S:/Users/deinesji/HPA/Salus/lema/uniqueExperiments'
comboDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/combinationGrids/spatialFilters'
load(file = paste0(comboDir,'/uniqueExperiments_gmd4p_Sheridan6_top7.RData'))

# gmd4+ template raster grid (based on CDL clipped to study region boundary)
aeaProj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
template <- raster(nrow = 4359, ncol = 7108, crs = aeaProj,
                    xmn = -522870, xmx = -309630, ymn = 1777860, 
                    ymx = 1908630) 
template[] <- 1:ncell(template)
templateVector <- getValues(template)

# aoi
gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries'
aoi <- readOGR(gisDir, 'Sheridan6_fromKGS_sd6', verbose = FALSE)
aoi <- spTransform(aoi, CRS(aeaProj))
```

# SALUS Maps and Summaries

## Load Data and get Expt Distributions
Results need to be distributed in space, but first, taken a quick glance at the unweighted results for the experiments run

### YIeld

```{r unweightedYieldResults, fig.width = 5, fig.height = 4.25}
# load formatted results output
expt.annual <- read_csv(paste0(resultsDir, '/', runname, '.csv'))

# get yield statistics by year/crop/irrigation status (not spatially weighted)
# species names
cropKey <- data.frame(SpeciesID = c('MZ', 'WH','SB','SG','AL'),
                      crop = c('CORN','WHEAT','SOYBEANS','SORGHUM','ALFALFA'))

uYieldStats <- expt.annual %>%
  filter(SpeciesID %in% cropKey$SpeciesID) %>%
  group_by(Year, SpeciesID, irrigation) %>%
  do(data.frame(t(quantile(.$GWAD, probs = c(.1, .25, .5, .75, .9),na.rm=TRUE)))) %>%
  left_join(cropKey)

# plot
ggplot(uYieldStats, aes(x=Year, y = X50., color = irrigation)) +
  geom_ribbon(aes(x = Year, ymin = X25., ymax = X75., fill=irrigation), alpha = .4) +
  geom_line() +
  facet_wrap(~crop) +
  ylab('Yield (kg/ha)') + xlab('') + 
  ggtitle('Median and IQR Yield, Lema Unweighted Experiments') +
  theme_bw() + theme(legend.position = 'bottom')#, legend.title = element_blank())
```

### Irrigation Water Used

```{r unweightedIrrResults, fig.height = 2.5, fig.width = 4}
# get summary statistics for all irrigated experiments
uIrrStats <- expt.annual %>%
  filter(irrigation == 'Y') %>%
  group_by(Year) %>%
  do(data.frame(t(quantile(.$IRRC_mm, probs = c(.1, .25, .5, .75, .9),na.rm=TRUE))))

# plot
ggplot(uIrrStats, aes(x=Year, y = X50.)) +
  geom_ribbon(aes(x = Year, ymin = X25., ymax = X75.), alpha = .4) +
  geom_line() +
  ylab('Irrigation (mm)') + xlab('') + 
  ggtitle('Median and IQR Irrigation Applied, Lema Unweighted Experiments') +
  theme_bw() 
```

### Precipitation

```{r unweightedPrecip, fig.height = 2.5, fig.width = 4}
uIPrecipStats <- expt.annual %>%
  group_by(Year) %>%
  do(data.frame(t(quantile(.$PREC_mm, probs = c(.1, .25, .5, .75, .9),na.rm=TRUE))))

ggplot(uIPrecipStats, aes(x=Year, y = X50.)) +
  geom_ribbon(aes(x = Year, ymin = X25., ymax = X75.), alpha = .4) +
  geom_line() +
  ylab('Precip (mm)') + xlab('') + 
  ggtitle('Median and IQR Precip, Lema Unweighted Experiments') +
  theme_bw() 
```


## Spatially Distribute
Distribute results based on grid cell ids. Rasters exported to a maps subfolder of the results directory

## function
Function to apply the spatialize function to all years, and export rasters

```{r funs_spatializeAll}
# returns a raster stack for value
spatializeAll <- function(expt.df, variable, yearsWanted, gridkey, template, aoi,
                          writeOut = 'N', outFolder = ''){
                  rasList <- list()
                  for(year in yearsWanted){
                    rasName <- as.character(year)
                    ras <- salustools::spatialize(expt.df, variable, year, 
                                                  gridkey, template,aoi)
                    if(writeOut == 'Y'){
                      fname <- paste0(outFolder, '/',variable, '_',year, '.tif')
                      writeRaster(ras, file = fname)
                    }
                    rasList[[rasName]] <- ras
                  }  
                  rasStack <- stack(rasList, quick = TRUE)
}
```


## Spatial maps

```{r globalVars}
# load formatted results output
expt.annual <- read_csv(paste0(resultsDir, '/', runname, '.csv'))

# create subfolder if needed
outdir <- paste0(resultsDir, '/', subf)
if(!dir.exists(outdir)){
  dir.create(file.path(resultsDir, subf), FALSE)
}

# vector of years wanted
yearsWanted <- startYear:endYear
```

### Irrigation

```{r spIrr, fig.width = 9, fig.height = 5}
irrStack <- spatializeAll(expt.annual, 'IRRC_mm', yearsWanted, gridExps,
                          template, aoi, writeOut = 'N', outFolder = outdir)
spplot(irrStack, main = 'SALUS Irrigation (mm)', maxpixels = 200000)
```

### Precip

```{r spPrecip, fig.width = 9, fig.height = 5}
pcpStack <- spatializeAll(expt.annual, 'PREC_mm', yearsWanted, gridExps,
                          template, aoi, writeOut = 'N', outFolder = outdir)
spplot(pcpStack, main = 'NLDAS Precip (mm)', maxpixels = 200000)
```

### Recharge

```{r spRecharge, fig.width = 9, fig.height = 5}
rchStack <- spatializeAll(expt.annual, 'DRNC_mm', yearsWanted, gridExps,
                          template, aoi, writeOut = 'N', outFolder = outdir)
spplot(rchStack, main = 'SALUS Recharge (mm)', maxpixels = 200000)


#save the stacks
save(irrStack, pcpStack, rchStack, file = paste0(outdir,
                                                 '/All_Irr_Rch_precip_stacks.RData'))
```

### Yield
Break down by crop and irrigation status

#### Irrigated

```{r spYield_irr, fig.width = 9, fig.height = 5}
yield <- 'GWAD'

crops <- c('MZ','WH','SG','SB','AL')

irrCrops <- expt.annual %>% filter(irrigation == 'Y') 

irrCropsList <- list()
for(m in 1:length(crops)){
  # subset results for crop
  crop <- crops[m]
  cropdf <- irrCrops %>% filter(SpeciesID == crop)
  cropStack <- spatializeAll(cropdf, yield, yearsWanted, gridExps,
                          template, aoi, writeOut = 'N', outFolder = outdir)
  irrCropsList[[crop]] <- cropStack
}

# save as RData
save(irrCropsList, file = paste0(outdir, '/All_IrrigatedCrops_YieldMaps.RData'))

# ie corn yield map
cornStacki <- irrCropsList[['MZ']]
spplot(cornStacki, main = 'Irrigated Corn Yield, kg/ha', maxpixels = 100000)
```

#### Not Irrigated

```{r spYield_rain, fig.width = 9, fig.height = 5}
yield <- 'GWAD'

crops <- c('MZ','WH','SG','SB','AL')

rainCrops <- expt.annual %>% filter(irrigation == 'N') 

rainCropsList <- list()
for(m in 1:length(crops)){
  # subset results for crop
  crop <- crops[m]
  cropdf <- rainCrops %>% filter(SpeciesID == crop)
  cropStack <- spatializeAll(cropdf, yield, yearsWanted, gridExps,
                          template, aoi, writeOut = 'N', outFolder = outdir)
  rainCropsList[[crop]] <- cropStack
}

# save as RData
save(rainCropsList, file = paste0(outdir, '/All_RainfedCrops_YieldMaps.RData'))

# ie corn yield map
cornStackr <- rainCropsList[['MZ']]
spplot(cornStackr, main = 'Rainfed Corn Yield, kg/ha', maxpixels = 100000)
```


## some individual year maps

```{r indvlMaps}
spplot(irrStack[['X2008']], main = '2008 salus irrigation (mm)')
spplot(irrStack[['X2012']], main = '2012 salus irrigation (mm)')
spplot(irrStack[['X2013']], main = '2013 salus irrigation (mm)')

spplot(rchStack[['X2008']], main = '2008 salus recharge (mm)')
spplot(rchStack[['X2012']], main = '2012 salus recharge (mm)')
spplot(rchStack[['X2013']], main = '2013 salus recharge (mm)')

spplot(cornStacki[['X2011']], main = '2011 Irrigated Corn Yield (kg/ha)')
spplot(cornStackr[['X2011']], main = '2011 Rainfed Corn Yield (kg/ha)')
```


## Data Table Synthesis

* precip is averaged over the domain and kept in mm
* irrigation includes total volumetric water, total area, and mean depth
* drainage...
* yields - summed by crop type

```{r waterVarsTable}
# get total irrigation over time
irrVolStack <- irrStack * 1e-3 * 30 * 30  # volumetric rasters in cubic meters

irrByYear <- data.frame(year = yearsWanted,
                        irr_m3 = cellStats(irrVolStack, stat = 'sum', na.rm=TRUE))
irrByYear$irr_km3 <- irrByYear$irr_m3 * 1e-6

# get total ppt  over time
pcpByYear <- data.frame(year = yearsWanted,
                        precip_mm = cellStats(pcpStack, stat = 'mean', na.rm=TRUE))

# get total recharge over time
rchVolStack <- rchStack * 1e-3 * 30 * 30  # volumetric rasters

rchByYear <- data.frame(year = yearsWanted,
                        rch_m3 = cellStats(rchVolStack, stat = 'sum', na.rm=TRUE))
rchByYear$rch_km3 <- rchByYear$rch_m3 * 1e-6

# combine
waterVars <- irrByYear %>%
  left_join(pcpByYear, by = 'year') %>%
  left_join(rchByYear, by = 'year')

waterVarsLong <- waterVars %>%
  select(c(year, irr_km3, precip_mm, rch_km3)) %>%
  gather(., key = variable, value = value, irr_km3:rch_km3)

ggplot(waterVarsLong, aes(x=year, y = value)) +
  geom_line() + 
  facet_wrap(~variable, nrow=3, scale = 'free_y') +
  theme_bw() +
  ylab('Total Recharge (km^3)       Annual Precip (mm)     Irrigation (km^3)') +
  ggtitle('Salus Water Variables')

# export
write.csv(waterVars, row.names = FALSE,
          file = paste0(resultsDir,'/WaterVars_meansTotals.csv'))
```


### Yield
I'm going to need to revisit these to get summary stats, not just mean!

#### Irrigated

```{r getYields}
crops <- c('MZ','WH','SG','SB','AL')

yielddf <- data.frame(year = rep(yearsWanted,2),
                      irrigated = rep(c('Y','N'), length(yearsWanted)),
                      MZ_kg = NA,
                      WH_kg = NA,
                      SG_kg = NA,
                      SB_kg = NA,
                      AL_kg = NA,
                      MZ_kgha = NA,
                      WH_kgha = NA,
                      SG_kgha = NA,
                      SB_kgha = NA,
                      AL_kgha = NA)

# get yield for Sheridan: rainfed
for(crop in crops) {
  # extract crop
  yieldStack <- rainCropsList[[crop]]
  # convert to kg per pixel (900m2) to be able to sum total
  yield30 <- yieldStack / 10000 * 900
  kgname <- paste0(crop,'_kg')
  yielddf[yielddf$irrigated=='N',kgname] <- cellStats(yield30, stat = 'sum',
                                      na.rm=TRUE)
  # get mean kg.ha  
  kghaname <- paste0(crop,'_kgha')
  yielddf[yielddf$irrigated=='N',kghaname] <- cellStats(yieldStack, stat = 'mean',
                                                        na.rm=TRUE)
}

# get  kg yield for Sheridan: irrigated
for(crop in crops) {
  # extract crop
  yieldStack <- irrCropsList[[crop]]
  # convert to kg per pixel (900m2) to be able to sum total
  yield30 <- yieldStack / 10000 * 900
  kgname <- paste0(crop,'_kg')
  yielddf[yielddf$irrigated=='Y',kgname] <- cellStats(yield30, stat = 'sum',
                                      na.rm=TRUE)
  # get mean kg.ha  
  kghaname <- paste0(crop,'_kgha')
  yielddf[yielddf$irrigated=='Y',kghaname] <- cellStats(yieldStack, stat = 'mean',
                                                        na.rm=TRUE)
}

# turn to long format
yieldlong <- yielddf %>%
  gather(key = crop, value = value, MZ_kg:AL_kgha) %>%
  mutate(units = substr(crop, start = 4, stop = nchar(crop)),
         crop = substr(crop, start = 1, stop = 2)) %>% 
         spread(key = units, value = value)

ggplot(yieldlong, aes(x=year, y = kgha, color=irrigated)) +
  geom_line() + 
  facet_wrap(~crop) + theme_bw() + 
  ggtitle('mean sheridan yield (kg/ha)')

ggplot(yieldlong, aes(x=year, y = kg, color=irrigated)) +
  geom_line() + 
  facet_wrap(~crop) + theme_bw() +
  ggtitle('total sheridan yield (kg)')

write.csv(yieldlong, row.names=FALSE,
          file = paste0(resultsDir,'/yields_meansTotals.csv'))
```
