==> code_raw_knitted/10.00_Presentation_Figures.Rmd <==
---
title: "Presentation Figures"
author: "Jill Deines"
date: "November 30, 2017"
output: 
  html_document:
    toc: yes
---

Goal: make figures for AGU 2017 presentation

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path='figure/10.00_PresFigs/',
                      cache = FALSE)
```

**R Packages Needed**

```{r packages, warning=FALSE, message=FALSE}
library(raster)
library(rgdal)
library(grid)
library(colorRamps)
library(fields)
library(RColorBrewer)
#library(latticeExtra)
library(viridis)
#library(rasterVis)

#library(devtools)
#install_github("jdeines/salustools")
library(salustools) # for nqblack ggplot theme

library(tidyverse)
library(lubridate)
```

**Directories**

```{r dirs}
gisBoundsDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/GIS/boundaries' 
```


# water use - wimas
what was the volume reduction between 2002-2012 and LEMA?


```{r wimasDrop, dpi = 400, fig.height = 3, fig.width = 4, dev=c('png','pdf')}
# load data
welldir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/wellData/WIMAS_AnthonyCleaned_JillFormatted'
wellsSheridan <- 'WIMAS_Sheridan_1990-2016.csv'
# get annual sums and means
wimas <- read_csv(paste0(welldir, '/', wellsSheridan))
wimasstats <- wimas %>%
  group_by(year) %>%
  summarise(irr_m3 = sum(volume_m3, na.rm=TRUE),
            totalArea_m2 = sum(area_m2, na.rm=TRUE),
            meanDepth_m = mean(depth_m, na.rm=TRUE)) %>%
  mutate(irrigated_km2 = totalArea_m2 * 1e-6,
         irr_km3 = irr_m3 * 1e-6,
         masterid = 'Sheridan6',
         datasource = 'WIMAS') %>%
  dplyr::select(-c(irr_m3, totalArea_m2, masterid)) %>%
  gather(.,key = variable, value = value, meanDepth_m:irr_km3)

# plots
annualIrrigation <- wimasstats %>% filter(variable == 'irr_km3' & year > 2001)
annualIrrigation$period <- c(rep('before',11),rep('lema',4))



ggplot(annualIrrigation,
       aes(x=year, y = value, group = period, color=period)) +
  geom_boxplot() + geom_point() + ylim(0,45) +
  theme_nqblack() + xlab("") + ylab('Total Irrigation Volume')+
  ggtitle('36% reduction')+
    theme(legend.position='none',
        axis.text=element_text(size=12),
        legend.text=element_text(size=10),
        axis.title=element_text(size=13),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())

# get difference
means <- annualIrrigation %>%
  group_by(period) %>%
  summarize(mean = mean(value))

# percent reduction
(1 - (means[means$period == 'lema','mean']/means[means$period == 'before','mean'])) * 100
  
```

###Add irrigation
I'm curious about the relative changes in irrigated area

```{r irrChange,dpi = 400, fig.height = 3, fig.width = 4, dev=c('png','pdf')}
tabDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/tabular'
cdl_sd6 <- 'CDL_SheridanTrace_2006-2016.csv'
aimName <- 'irrArea_climate_from_AIMRRB.csv'

aim <- read_csv(paste0(tabDir,'/', aimName)) %>%
  filter(masterid == 'sheridan6') %>%
  dplyr::select(c(year,irrigated_km2)) %>%
  mutate(datasource = 'AIM') %>%
  gather(., key = variable, value = value, irrigated_km2) %>%
  filter(year > 2001 & year != 2006) %>%
  mutate(period = c(rep('before',10),rep('lema',4)))

ggplot(aim,
       aes(x=year, y = value, group = period, color=period)) +
  geom_boxplot() + geom_point() + 
  theme_nqblack() + xlab("") + ylab('Total Irrigation Area')+
  ggtitle('')+ ylim(0, 120) + ggtitle('9% increase') +
    theme(legend.position='none',
        axis.text=element_text(size=12),
        legend.text=element_text(size=10),
        axis.title=element_text(size=13),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())

# get difference
means <- aim %>%
  group_by(period) %>%
  summarize(mean = mean(value))

# percent reduction
(1 - (means[means$period == 'lema','mean']/means[means$period == 'before','mean'])) * 100

```



# Water Levels

## Sheridan Water Level Changes
See 01.60_gwLevels.Rmd for figures

```{r process_waterlevels, eval = FALSE}
wlDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/waterLevels_Haacker'

# load just Sheridan and the GMD with hole
masterDF <- read_csv(paste0(wlDir,'/20171209_Combined_wt_vol.csv')) %>%
  #filter(region != 'GMD4') %>%
  filter(Year > 2001) 
```

# SALUS Output

## irrigation volume

### salus vs wimas

```{r salus_vs_wimas_2007, dpi = 400, fig.height = 3, fig.width = 4, dev=c('png','pdf')}

resultsDir2 <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/experiments/5_LEMA_hist_fixed_v03/results'
irrigationSummary <- read_csv(paste0(resultsDir2,                                     '/irrigationSummary_salus_wimas_aim.csv'))


irrwater <- irrigationSummary %>% filter(variable == 'irr_km3' & year >= 2007)

ggplot(irrwater, aes(x = year, y = value, color=datasource)) +
  geom_line(lwd=1.2) + ylab('Irrigation Volume (km^3)') + xlab('') +
  geom_vline(xintercept=2013, col='goldenrod2', lty='dashed', lwd=1) +
  theme_nqblack() + 
  theme(legend.position='none',
        axis.text=element_text(size=12),
        legend.text=element_text(size=10),
        axis.title=element_text(size=13),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())
```

## Irrigation Maps by Year

```{r irrMaps_nqblack_2007, cache=TRUE, fig.width = 10, fig.height = 5, dpi = 400}
# load irr rasters
resultsDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/salus/experiments/5_LEMA_hist_fixed_v03/results/maps'
load(paste0(resultsDir, '/All_Irr_Rch_precip_stacks.RData'))

irrStack_2007 <- irrStack[[2:11]]

# reproject rasters?
classProj <- '+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-101 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0'
irrStack.proj <- projectRaster(irrStack_2007, crs = CRS(classProj))

# load Sheridan Lema bound (for under plotting - fill roads at "0")
lema <- readOGR(gisBoundsDir, 'Sheridan6_fromKGS_sd6', verbose = FALSE)
lema.proj <- spTransform(lema, CRS(proj4string(irrStack.proj)))
#lema.proj <- spTransform(lema, CRS(classProj))

# plot specs
pal <- bpy.colors(9)
labels <- as.character(2007:2016)
cacheDummy <- 8

# plot
spplot(irrStack.proj, col.regions = bpy.colors, col='gray29',
       names.attr=labels, maxpixels = 200000,layout = c(5,2),
       par.settings = list(panel.background=list(col="#262626"),
                           layout.widths=list(right.padding=6))) +
  latticeExtra::layer(sp.polygons(lema.proj, fill = pal[1], col=NA), under=T)

# legend title
grid.text('Irrigation Depth (mm)',x=unit(0.96, "npc"),y=unit(0.5, 'npc'), rot=-90)

# separate legend
# color ramp
# image.plot(z=0:18,add=F,legend.only=T,  col=c(backgroundcolor,matlab), 
#            horizontal = T,
#           smallplot=c(0.05,.95,.15,.2), 
#          axis.args=list(at=c(0,3,6,9,12,15,18), 
#                           labels=as.character(c(0,3,6,9,12,15,18))))

```

```{r irrMaps_gray_2007, cache = TRUE, fig.width = 10, fig.height = 5, dpi = 400, background='transparent'}

pal <- bpy.colors(9)
cacheDummy <- 6
# plot
spplot(irrStack.proj, col.regions = bpy.colors, col='gray29',
       names.attr=labels, maxpixels = 200000, layout = c(5,2),
       par.settings = list(panel.background=list(col="gray65"),
                           layout.widths=list(right.padding=6))) +
  latticeExtra::layer(sp.polygons(lema.proj, fill = pal[1], col=NA), under=T)

# legend title
grid.text('Irrigation Depth (mm)',x=unit(0.96, "npc"),y=unit(0.5, 'npc'), rot=-90)
```


```{r irrMaps_gray_2007_3r, cache = TRUE, fig.width = 9.8, fig.height = 5.2, dpi = 400, background='transparent'}

pal <- bpy.colors(9)
cacheDummy <- 6
# plot
spplot(irrStack.proj, col.regions = bpy.colors, col='gray29',
       names.attr=labels, maxpixels = 200000, 
       par.settings = list(panel.background=list(col="gray65"),
                           layout.widths=list(right.padding=6))) +
  latticeExtra::layer(sp.polygons(lema.proj, fill = pal[1], col=NA), under=T)

# legend title
grid.text('Irrigation Depth (mm)',x=unit(0.96, "npc"),y=unit(0.5, 'npc'), rot=-90)
```



## Lema WT changes
Change in LEMA water table prior to LEMA (1996-2012)

```{r lemaWT}
wlDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/waterLevels_Haacker'
wt <- raster(paste0(wlDir,'/','WTelevDiff_2012-1996_clip.tif'))
proj4string(wt)

spplot(wt, draw=T)
```


## HPA Depletion

```{r hpa_depletion, eval=FALSE}
wlDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/LEMAs/data/waterLevels_Haacker'
decline <- raster(paste0(wlDir,'/percent_decline_3m_sat_thick_set_100_clip.img'))

# col ramps
# blue and red color ramps pulled from color brewer (no white)
redRamp2 <- colorRampPalette(c('#fddbc7','#f4a582','#d6604d','#b2182b','#67001f'))
blueRamp2 <- colorRampPalette(rev(c('#d1e5f0','#92c5de','#4393c3','#2166ac','#053061')))

colorbreaks <- 200
biPal <- c(blueRamp2(colorbreaks/2),redRamp2(colorbreaks/2))

# scale range
palMin <- -.6
palMax <- .6 

spplot(decline, col.regions=(biPal))

```
